Chelsey's fixes to the quiz
I switched and used java for the quiz so that the questions would change dynamically. The issue is that it is not storing what the user puts in. That is because I am clearing the div each time you hit next so that all the questions don't show at once. I think an if statement that says hold the index of the answers would work but I don't know if the bubbles would still be filled in.
- Also I fixed the functionality of the buttons
-I still need to fix the progress bar at the top.